requirements.txt
black
matplotlib
numpy
openslide-python
pandas
pillow
pypubsub
scikit-image
seaborn
wxpython